Refactoring to remove all type based event timers.
