+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/shared/GeneralInfo.kt
@@ -1,4 +1,4 @@
-package com.jordangellatly.hermosahappyhour.ui.detail.info
+package com.jordangellatly.hermosahappyhour.ui.detail.shared
 
 import android.content.Intent
 import android.net.Uri
@@ -28,7 +28,6 @@ import com.jordangellatly.hermosahappyhour.model.Location
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.tower12
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourDivider
-import com.jordangellatly.hermosahappyhour.ui.detail.popup.HoursPopup
 import com.jordangellatly.hermosahappyhour.ui.home.getCurrentDateTime
 import com.jordangellatly.hermosahappyhour.ui.home.toString
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
