+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/special/SpecialEvent.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.unit.dp
 import com.jordangellatly.hermosahappyhour.model.Deal
 import com.jordangellatly.hermosahappyhour.model.sundaySilentDiscoSunset
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourDivider
+import com.jordangellatly.hermosahappyhour.ui.detail.shared.EventTimer
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 
 @Composable
@@ -26,7 +27,7 @@ fun SpecialEvent(
     specials: List<Deal>
 ) {
     Column(modifier = Modifier.padding(8.dp)) {
-        SpecialEventTimer(
+        EventTimer(
             weeklyHoursDescription = weeklyHoursDescription,
             eventStart = eventStart,
             eventEnd = eventEnd,
